Get highscores per skin in the front-end
